a few fixes, change some default settings
git-svn-id: http://svn.airdcpp.net/airsvn/trunk@243 ec3e54a6-43b8-4a3e-9c49-2a733a73d9fe

+++ Util.cpp
@@ -1281,8 +1281,8 @@ tstring Util::getDir(tstring dir) {
 			if(dir[dir.size() -1] == '\\')
 				directory = directory.substr(0, directory.size()-1);
 
-			int dpos = directory.rfind("\\");
-			if(dpos != (int)tstring::npos) {
+			size_t dpos = directory.rfind("\\");
+			if(dpos != tstring::npos) {
 				directory = directory.substr(dpos+1,directory.size());
 			}
 		}
@@ -1298,8 +1298,8 @@ tstring Util::validateDir(tstring dir) {
 			if (regex_match(Text::toT(directory), reg)) {
 				if(dir[dir.size() -1] == '\\')
 					directory = directory.substr(0, directory.size()-1);
-				int dpos = directory.rfind("\\");
-				if(dpos != (int)tstring::npos) {
+				size_t dpos = directory.rfind("\\");
+				if(dpos != tstring::npos) {
 					directory = directory.substr(0,dpos+1);
 				}
 			} else {
